+++ Source/SwiftLintFramework/Models/Configuration.swift
@@ -151,6 +151,7 @@ public struct Configuration: Equatable {
 
     public static func rulesFromYAML(yaml: Yaml? = nil) -> [Rule] {
         return [
+            ConditionalBindingCascadeRule(),
             ClosingBraceRule(),
             ColonRule(),
             CommaRule(),